Update the TodoService to only fetch todos for the current user

+++ src/Services/TodoServices/TodoService.cs
@@ -22,7 +22,10 @@ namespace todoProject.Services.TodoServices
 
         public async Task<TodoListDto> CreateTodoAsync(TodoListDto todo)
         {
+            var currentUser = await _userResolver.GetCurrentUserAsync();
+
             var todoToCreate = _mapper.Map<Todo>(todo);
+            todoToCreate.Owner = currentUser;
 
             await _context.AddAsync(todoToCreate);
             await _context.SaveChangesAsync();
@@ -32,7 +35,10 @@ namespace todoProject.Services.TodoServices
 
         public async Task<TodoListDto[]> GetAllTodosAsync()
         {
+            var currentUser = await _userResolver.GetCurrentUserAsync();
+
             var todos = await _context.Todos
+                .Where(t => t.Owner.Id == currentUser.Id)
                 .ToArrayAsync();
             
             return _mapper.Map<TodoListDto[]>(todos);
@@ -40,17 +46,32 @@ namespace todoProject.Services.TodoServices
 
         public async Task<TodoListDto> GetById(int id)
         {
+            var currentUser = await _userResolver.GetCurrentUserAsync();
             var todo = await _context.Todos
                 .FindAsync(id);
+            
+            if (todo.Owner.Id != currentUser.Id) 
+            {
+                // TODO: Throw an exception about not having access or something
+                return null;
+            }
             
             return _mapper.Map<TodoListDto>(todo);
         }
 
         public async Task DeleteTodoAsync(int id)
         {
+            var currentUser = await _userResolver.GetCurrentUserAsync();
+
             var todo = await _context.Todos.FindAsync(id);
 
             if (todo == null) return;
+
+            if (todo.Owner.Id != currentUser.Id)
+            {
+                // TODO: Throw an exception about not having access or something
+                return;
+            }
 
             _context.Todos.Remove(todo);
             await _context.SaveChangesAsync();
@@ -58,9 +79,16 @@ namespace todoProject.Services.TodoServices
 
         public async Task<TodoListDto> UpdateTodoAsync(TodoListDto updatedTodo)
         {
+            var currentUser = await _userResolver.GetCurrentUserAsync();
             var originalTodo = await _context.Todos.FindAsync(updatedTodo.Id);
 
             if (originalTodo == null) return null;
+
+            if (originalTodo.Owner.Id != currentUser.Id)
+            {
+                // TODO: Throw an exception about not having access or something
+                return null;
+            }
 
             originalTodo.Title = updatedTodo.Title;
             originalTodo.Done  = updatedTodo.Completed;
